AdventureScene

+++ DANGER/Assets/Luis/ControlSelectedPlayer.cs
@@ -32,8 +32,8 @@ public class ControlSelectedPlayer : MonoBehaviour
         if(player != null)
         {
             Debug.Log(player.GetComponent<Sensor>().playerHealth);
-            playerHealthSlider.value = player.GetComponent<Sensor>().playerHealth / 100.0f;
-            palyerStressSlider.value = player.GetComponent<Sensor>().playerStress / 100.0f;
+            playerHealthSlider.value = (float)player.GetComponent<Sensor>().playerHealth / 100.0f;
+            palyerStressSlider.value = (float)player.GetComponent<Sensor>().playerStress / 100.0f;
         }
     }
